Todavía falta corregir crashes

+++ BusFinder/app/src/main/java/com/gammasoft/busfinder/view/fragment/borrar/TarjetaChofer.kt
@@ -4,6 +4,7 @@ import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import android.widget.Toast
 import androidx.annotation.AnimRes
 import com.gammasoft.busfinder.R
 import com.gammasoft.busfinder.databinding.TarjetaBorrarChoferBinding
@@ -24,7 +25,7 @@ class TarjetaChofer(private val localDB: Crud,
     private val binding get() = _binding!!
 
     fun mostrar(@AnimRes enterAnim: Int = R.anim.zoom_in,
-                @AnimRes exitAnim: Int = R.anim.zoom_out) = TarjetaChofer(localDB, chofer).withEnterAnim(enterAnim).withExitAnim(exitAnim)
+                @AnimRes exitAnim: Int = R.anim.zoom_out) = this.withEnterAnim(enterAnim).withExitAnim(exitAnim)
 
     override fun onCreateView(
         inflater: LayoutInflater,
@@ -48,8 +49,16 @@ class TarjetaChofer(private val localDB: Crud,
 
         binding.btnBorrar.setOnClickListener{
             CoroutineScope(Dispatchers.IO).launch{
-                localDB.deleteChofer(chofer)
-                CloudDataBase.delete("Chofer", chofer.getUsuario())
+                localDB.getChoferByRFC(chofer.getRfc()).observe(viewLifecycleOwner){
+                    if(it.getNombre() == chofer.getNombre()){
+                        localDB.deleteChofer(chofer)
+                        chofer.setAdministrador("")
+                        CloudDataBase.addChofer(chofer)
+
+                        Toast.makeText(requireContext(), "¡Chofer borrado con éxito!", Toast.LENGTH_SHORT).show()
+                        dismiss()
+                    }
+                }
             }
         }
     }
